requirements.txt
streamlit
openai>=1.0.0
pandas
fpdf
pytesseract
pillow
python-dotenv
streamlit-webrtc
speechrecognition
pyaudio
numpy
av
